0.2 - File structure refactoring
