+++ backend/handlers/__init__.py
@@ -1,0 +1,1 @@
+from backend.handlers import *
