+++ backend/models/BaseNode.py
