--- core/TreeNode.py
@@ -1,6 +1,0 @@
-class TreeNode(object):
-
-    def __init__(self, text, values, icon=None):
-        self.text = text
-        self.values = values
-        self.icon = icon
