+++ gui/forms/AddStoryItemForm.py
