+++ gui/forms/RemoveItemForm.py
