+++ gui/forms/UpdateItemForm.py
@@ -1,54 +1,54 @@
-from uuid import uuid4
-
 import PySimpleGUI as sg
 
 from common.constants import *
 from gui.layout import button_params
 
 
-class AddExampleForm(object):
-    FORM_NAME = 'Add example'
-    INPUT_LABEL = 'Example'
+class UpdateItemForm(object):
     EMPTY_TEXT_ERROR = "Text cannot be empty."
+    INPUT_KEY = 'updated_text'
     ITEM_WITH_KEY_EXISTS_ERROR = "Item with such name/text already exists."
-    INPUT_KEY = 'example_text'
 
-    def __init__(self, return_to, parent_name, handler, tree):
+    def __init__(self, return_to, form_name, updated_item_key, updated_item_type, handler, tree, stories, stories_tree):
         self.return_to = return_to
-        self.parent_name = parent_name
+        self.updated_item_key = updated_item_key
+        self.updated_item_type = updated_item_type
         self.handler = handler
+        self.stories = stories
         self.tree = tree
-        self.form = sg.FlexForm(self.FORM_NAME, return_keyboard_events=True).layout(self.layout())
+        self.stories_tree = stories_tree
+        self.form = sg.FlexForm(form_name, return_keyboard_events=True)
 
-    def layout(self):
-        layout = [
-            [sg.Text(self.INPUT_LABEL), sg.InputText(key=self.INPUT_KEY)],
+    def layout(self, old_text):
+        update_text_layout = [
+            [sg.InputText(key=self.INPUT_KEY, default_text=old_text)],
             [sg.Button(ACTION_SUBMIT, **button_params(green_button)),
              sg.Button(ACTION_CANCEL, **button_params(orange_button))]
         ]
-        return layout
+        return update_text_layout
 
     def validate(self, values):
         error = None
         if values[self.INPUT_KEY] == '':
             error = self.EMPTY_TEXT_ERROR
-        if values[self.INPUT_KEY] in self.handler.items:
+        elif values[self.INPUT_KEY] in self.handler.items:
             error = self.ITEM_WITH_KEY_EXISTS_ERROR
         return error
 
     def process(self):
         self.return_to.Disable()
-        new_item = None
+        self.form.Layout(self.layout(self.updated_item_key))
+
         while True:
             button, values = self.form.Read()
             if button in DEFAULT_FORM_SUBMIT_ACTIONS:
                 error = self.validate(values)
                 if not error:
-                    new_item = values[self.INPUT_KEY]
-                    self.handler.add_example_to_node(self.parent_name, new_item)
-                    self.tree.Update(values=self.handler.export_to_pysg_tree())
-                    self.tree.see(new_item)
-                    self.tree.selection_set([new_item])
+                    self.handler.update_node_value(self.updated_item_key, values['updated_text'])
+                    self.tree.Update(self.handler.export_to_pysg_tree())
+                    self.tree.see(values['updated_text'])
+                    self.tree.selection_set([values['updated_text']])
+                    self.stories_tree.Update(self.stories.export_to_pysg_tree())
                 else:
                     sg.Popup(error, icon=sg.SYSTEM_TRAY_MESSAGE_ICON_CRITICAL, keep_on_top=True)
                 self.form.close()
@@ -56,7 +56,5 @@ class AddExampleForm(object):
             elif button in DEFAULT_FORM_CANCEL_ACTIONS:
                 self.form.close()
                 break
-        # self.handler.sort_alphabetically()
         self.return_to.bring_to_front()
         self.return_to.Enable()
-        return new_item
